refactor: extract validation schema and add payload configuration fields for DataStream
