+++ src/utils/constants.js
@@ -3,3 +3,5 @@ export const CDN_MAXIMUM_TTL_MIN_VALUE = 3
 export const TTL_MAX_VALUE_RECORDS = 604800
 export const TTL_DEFAULT = 3600
 export const TTL_DEFAULT_ANAME = 20
+export const MIN_PAYLOAD_SIZE_IN_BYTES = 1000000
+export const MAX_PAYLOAD_SIZE_IN_BYTES = 2147483647
